feat(update liquids page): update liquids page
update liquids page

close AUTH-1828
